requirements.txt
pypugjs==5.8.1
python-dotenv==0.7.1
django-rosetta==0.9.3
mutagen==1.42.0
tweepy==3.7.0
Django==2.2.2
fuzzywuzzy==0.17.0
python-levenshtein==0.12.0
setuptools==41.0.1
html2text==2018.1.9
markdown==3.1.1
colorgram==1.2.0
pillow==6.0.0